refactor(pipeline/pii): extract AuthCredentials/ErrorLog/Url/JsonBody + populate Facade
Phase 6 commit 5/7. Splits the remaining strategies out of PiiRedactor.ts:

- AuthCredentials.ts: redactToken/Otp/Cookie + looksLikeToken/Otp/Cookie value sniffers + shared redactOpaqueSecret.

- ErrorLog.ts: redactErrorMessage + redactSensitiveEnum + SENSITIVE_SCRAPER_ENUMS (only module allowed to import ScraperErrorTypes).

- Facade.ts (populated): PATH_TAIL_TO_CATEGORY routing, classifyKey, createCensorFn, dispatchStrategy, STRING_STRATEGIES, unified redact() entry point with auth-first dispatch.

- Url.ts: redactUrl / redactUrlFull + PII_QUERY_KEYS + tryParseUrl + path-segment masking.

- JsonBody.ts: redactJsonBody walker (redactNode/Object/Array/Leaf) + fallback regex sweep.

PiiRedactor.ts is now a re-export shim — every public symbol preserved byte-identical via per-module re-exports. 138/138 PiiRedactor tests pass.
